created basic layout
