--- src/Components/StyleComponents/MainLayout.jsx
@@ -1,18 +1,0 @@
-import React from "react";
-import CssBaseline from "@material-ui/core/CssBaseline";
-import Typography from "@material-ui/core/Typography";
-import Container from "@material-ui/core/Container";
-
-export default function SimpleContainer() {
-  return (
-    <React.Fragment>
-      <CssBaseline />
-      <Container maxWidth="xl" style={{ padding: "0" }}>
-        <Typography
-          component="div"
-          style={{ backgroundColor: "#cfe8fc", height: "100vh" }}
-        />
-      </Container>
-    </React.Fragment>
-  );
-}
