+++ src/Components/main.module.css
@@ -1,0 +1,3 @@
+.container {
+    height: 100%;
+}
